backend/config/schema.py
import graphene
import graphql_jwt


class Query(graphene.ObjectType):
    title = graphene.String(default_value="Easy Commerce!")


class Mutation(graphene.ObjectType):
    token_auth = graphql_jwt.ObtainJSONWebToken.Field()
    verify_token = graphql_jwt.Verify.Field()
    refresh_token = graphql_jwt.Refresh.Field()


schema = graphene.Schema(mutation=Mutation, query=Query)
