pyproject.toml
[tool.poetry]
name = "easy-commerce-backend"
version = "0.0.1"
description = "Easy Commerce Single Vendor E-Commerce Website back-end created with Django and GraphQL"
authors = ["Shakib Hossain <shakib609@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.5"
environs = "^9.3.1"
psycopg2-binary = "^2.8.6"
graphene-django = "^2.15.0"
django-graphql-jwt = "^0.3.1"
PyJWT = "<2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
